requirements.txt
python-dotenv==1.2.1
websockets==16.0
